Enable sign in persistence and sign out. Update navbar.

+++ frontend/src/Contexts/UserContext.tsx
@@ -1,10 +1,12 @@
 import { User } from '@firebase/auth';
 import React, { Dispatch, HTMLAttributes, SetStateAction, useContext, useState } from 'react';
+import { useEffect } from 'react';
+import { auth } from '../Services/FirebaseService';
 
 type UserContextType = {
   readonly user: User | undefined;
   readonly isLoggedIn: boolean;
-  performLogIn: (user: User) => void;
+  performLogIn: (user: User, keepLoggedIn: boolean) => void;
   performLogOut: () => void;
   setUser: Dispatch<SetStateAction<User | undefined>>;
 };
@@ -15,15 +17,40 @@ export const UserContextProvider = ({ children }: HTMLAttributes<HTMLElement>) =
   const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
   const [user, setUser] = useState<User | undefined>();
 
-  const performLogIn = (user: User) => {
+  const handleReload = (usr: User | undefined | null) => {
+    if (usr) {
+      setUser(usr);
+      setIsLoggedIn(true);
+    } else {
+      localStorage.clear();
+    }
+  };
+
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged(usr => {
+      handleReload(usr);
+    });
+
+    return unsubscribe();
+  }, []);
+
+  const performLogIn = (user: User, keepLoggedIn: boolean) => {
+    keepLoggedIn && localStorage.setItem('user', JSON.stringify(user));
     setUser(user);
     setIsLoggedIn(true);
     //Create Session In Mongo
   };
 
-  const performLogOut = () => {
-    localStorage.clear();
-    setIsLoggedIn(false);
+  const performLogOut = async () => {
+    auth
+      .signOut()
+      .then(() => {
+        localStorage.clear();
+        setIsLoggedIn(false);
+      })
+      .catch(() => {
+        alert('Error! Could Not Log Out. Please Try Again!');
+      });
   };
 
   return (
